fix inputChaining
