+++ telegram-bot/src/main/kotlin/eu/vendeli/tgbot/core/TelegramUpdateHandler.kt
@@ -21,6 +21,7 @@ import eu.vendeli.tgbot.utils.NewCoroutineContext
 import eu.vendeli.tgbot.utils.checkIsLimited
 import eu.vendeli.tgbot.utils.findAction
 import eu.vendeli.tgbot.utils.invokeSuspend
+import eu.vendeli.tgbot.utils.process
 import eu.vendeli.tgbot.utils.processUpdate
 import kotlinx.coroutines.cancelChildren
 import kotlinx.coroutines.channels.Channel
