+++ telegram-bot/src/main/kotlin/eu/vendeli/tgbot/types/internal/SingleInputChain.kt
@@ -3,7 +3,7 @@ package eu.vendeli.tgbot.types.internal
 import eu.vendeli.tgbot.types.internal.configuration.RateLimits
 import eu.vendeli.tgbot.utils.OnInputAction
 
-class SingleInputChain(
+data class SingleInputChain(
     internal val id: String,
     internal val inputAction: OnInputAction,
     internal val rateLimits: RateLimits,
